Remove extra alloc node stack push.
No longer need a separate stack to return an allocation. Use only the
allocation stack. Also, every region will always have an associated node
so no need to has a special node for region allocations, return the node
with a region. Both the region and the node will be used.